Add FIPS140-3 support for CLI commands
Add support for an environment variables that allows all CLI commands to run with the various FIPS options set. The variable can be supplied just the same as any other environment variable that is picked up for a server, client or tool. For Server and Client, the value of the property is retrieved from the file specified, allowing for the use of a singular property. Tools generated from the tool/tool.bat script only need the property to set for them to add the relevant FIPS flags.

New subrountines have been added for Server and Client scripts, while for all other tools, it is part of their setup steps. For Client and Server FIPS is only checked on relevant operations e.g. `./server start testServer`, but not on `./server create testServer` which does not need to apply FIPS anyway. There is no similar mechanism of being able to choose what a tool does as all command line options are passed to the jar to operate on.

If the value of `ENABLE_FIPS140_3` is `false` then FIPS is disabled.
Empty String e.e. ENABLE_FIPS140_3="" is set for either Windows or *nix then these will work with IBM SDK8
On Windows if JAVA_HOME is set to an Java 8 JVM, it is updated to be `!JAVA_HOME!\jre`. so the JAVA_HOME FIPS 140-3 IBM Java 8 check must not include `\jre` in the path as it is already set

The scripts based on which property denotes which Java executable is to be run (JAVE_HOME, JRE_HOME or System Java) determine whether to use the IBM SDK 8 or IBM Semeru settings. the user cannot override this without changing the version of Java being used.
If it is found that the level of Java does not support FIPS e.g. IBM Semeru 8 or fixpacks that do not support the level of java required, FIPS will not be enabled.

For IBM SDK 8, the presence of the `fips140-3` directory in the Java install location is used.
For IBM Semeru, the contents of the java.security file is checked to see if it includes the `OpenJCEPlusFIPS.FIPS140-3-Strongly-Enforced` profile, which the Liberty profile extends. So earlier Semeru Releases which support FIPS with either the strict or weak profiles are not enabled via the commands

For the Semeru Profile, the new `-Djava.security.propertiesList` argument is used, which allows the specifying of multiple profiles that are expected to be linked to each other. One of these profiles should extend the `FIPS140-3-Liberty` profile.
The Liberty profile is prepended to the value of the file path. if the value of the property is "", then the base Liberty profile will be used.

From the provided profile files, the last profile in the list is used to determine the value of `-Dsemeru.customProfile`.

It is still possible for a user to supply the jvm.options for FIPS which would