When delegating to a parentFirst loader check parent before defineClass
When delegating to a loader we excludeParent search.  If the delegated
to loader is parentFirst, but the delegating loader has not searched
parent yet then we need to do an extra check for parent before
defining the class.  Otherwise we risk defining a class in the
parentFirst loader which should come from the parent.