MCP: Create special arguments in one place
Previously we had a two-step process for creating the values for special
arguments before calling a tool.

After parsing the arguments from the request, we would then insert a
placeholder for every special argument. Later, in McpServlet.callTool,
we would have those placeholders replaced by the actual values.

Instead of replacing placeholders, we can use the tool metadata to find
which parameters need special arguments and what types they should be so
that all the processing can be done in one place.