Add tests for FIPS 140-3 Enablement and securityUtility
Add tests that just cover the server and client scripts taking the ENABLE_FIPS140_3 envvar and use it to ensure we can isolate those scripts behaviour.
Add tests that use securityUtility to set various enablement configurations and scenarios. includes starting servers and clients for end-to-end scenarios

Applies to both Semeru versions that support Liberty's minimum FIPS levels and IBM SDK 8

Checks within the test ensure we have a valid environment to run the tests. AlwaysPassesTest is included from when base requirements are not met

Update LibertyClient to support the ability to add the ability to addEnvVars as we do for LibertyServer.
Fix LibertyClient tracking of isStarted behaviour. isStopped was never used, but isStarted was never set back to false, once the client had stopped.