Prototype support for multiple FATSuites in a gradle project
- Add build.example.multi_fat to show the prototype in action where
suite.name and tested.features are defined in sub bundles.
tested.features can be defined in the bnd.bnd file like normal as well,
but this allows for each suite to define its own set of features it
tests.
- Support both running in gradle locally for a developer and in the
build systems.  A developer must set test.bucket.class in order to
designate which FATSuite to run if there are multiple of them.
- Update processing of fat-metadata.json to look for a suite specific
file when running in a local development gradle environment.