MCP: Support 2025-03-26
Some older clients we want to demo with support this older version of
the spec. The main incompatibility is the use of the
MCP-Protocol-Version header which isn't required.

The version in use is stored in the McpTransport so other classes can
query it if different behaviour is required for different versions.