Remove -ea flag from MCP test server
-ea means "enable assertions", and causes all asserts in the code to be
evaluated. At the moment, the netty feature has some issues which cause
asserts to fail. We need to add this flag so that our tests can run when
the netty feature is enabled.

We don't have any assert instructions in our tests or code any more, so
removing the flag does not affect our testing in any way. JUnit assert
methods in our FAT tests are still evaluated without this flag.